\name{GridLMM-package}
\alias{GridLMM-package}
\alias{GridLMM}
\docType{package}
\title{
\packageTitle{GridLMM}
}
\description{
\packageDescription{GridLMM}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{GridLMM}
\packageIndices{GridLMM}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{GridLMM}

Maintainer: \packageMaintainer{GridLMM}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
