% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GridLMM_GWAS_fast.R
\name{GridLMM_GWAS_fast}
\alias{GridLMM_GWAS_fast}
\title{GridLMM GWAS with adaptive algorithm.}
\usage{
GridLMM_GWAS_fast(formula, test_formula, reduced_formula, data, Y = NULL,
  weights = NULL, X, X_ID = "ID", centerX = FALSE, scaleX = FALSE,
  fillNAX = FALSE, X_map = NULL, relmat = NULL, h2_step = 0.01,
  h2_start = NULL, h2_start_tolerance = 0.001, max_steps = 100,
  method = "ML", inv_prior_X = NULL, target_prob = 0.99,
  proximal_matrix = NULL, downdate_Xs = NULL, V_setup = NULL,
  save_V_folder = NULL, diagonalize = T, svd_K = T, drop0_tol = 1e-10,
  mc.cores = my_detectCores(), clusterType = "mclapply", verbose = T)
}
\arguments{
\item{formula}{A two-sided linear formula as used in \code{\link[lme4]{lmer}} describing the fixed-effects
and random-effects of the model on the RHS and the response on the LHS. Note: correlated random-effects
are not implemented, so using one or two vertical bars (\code{|}) or one is identical. At least one random effect is needed.}

\item{test_formula}{One-sided formula for the alternative model (ML or BF), or full model (REML) to be applied to each test (ie \emph{marker}, or column of \code{X}). 
Each term on the RHS will be multiplied by a column of X to form a new covariate. 
Ex. \code{~1} specifices an intercept for each marker.
\code{~1+cov} species an intercept and slope on \code{cov} for each marker.}

\item{reduced_formula}{One-sided formula for the reduced model. Same format as \code{test_formula}. Should have fewer degrees of freedom than \code{test_formula}. Not used for REML models.}

\item{data}{A data frame containing the variables named in \code{formula}.}

\item{Y}{Optional matrix of responses. Currently only one response is evaluated.}

\item{weights}{An optional vector of observation-specific weights.}

\item{X}{Matrix of markers with \eqn{p} columns. Each column of X is used as a separate association test. Should have row names that correspond to the \code{X_ID} column of \code{data}.}

\item{X_ID}{Column of \code{data} that identifies the row of \code{X} that corresponding to each observation. 
It is possible that multiple observations reference the same row of \code{X}.}

\item{centerX, scaleX, fillNAX}{TRUE/FALSE for each. Applied to the \code{X} matrix before using \code{X} to form any GRMs.}

\item{X_map}{Optional. Data frame with information on each marker such as chromosome, position, etc. Will be appended to the results}

\item{relmat}{A list of matrices that are proportional to the (within) covariance structures of the group level effects. 
The names of the matrices should correspond to the columns in \code{data} that are used as grouping factors. All levels
of the grouping factor should appear as rownames of the corresponding matrix.}

\item{h2_step}{Step size of the grid}

\item{h2_start}{Optional. Matrix with each row a vector of \eqn{h^2} parameters defining starting values for the grid. Typically ML/REML solutions for the null model. 
If null, will be calculated using \link{GridLMM_ML}.}

\item{h2_start_tolerance}{Optional. Grid size for \link{GridLMM_ML} in finding ML/REML solutions for the mull model.}

\item{max_steps}{Maximum iterations of the heuristic algorithm per marker.}

\item{method}{One of 'ML', 'REML', or 'BF'. 'ML' implies Maximum Likelihood evaluation, with the LRT. 'REML' wimplies a Wald-test. 'BF' does posterior evaluation and calculates Bayes Factors.}

\item{inv_prior_X}{Vector of values for the prior precision of each of the fixed effects (including an intercept). Will be recycled if necessary.}

\item{target_prob}{See \strong{Details}.}

\item{proximal_matrix}{A matrix of size \eqn{p \times p} used to determine which markers should be dropped from the GRM when testing each marker. 
Each row is a vector of 0's or 1's, with element \code{(i,j)==1} if marker j should be dropped when testing marker i.}

\item{downdate_Xs}{A list providing information for updating GRMs based on proximal markers. 
Created based on \code{proximal_matrix} by this function, and can be re-passed to the function in future runs.}

\item{V_setup}{Optional. A list produced by a GridLMM function containing the pre-processed V decompositions for each grid vertex, 
or the information necessary to create this. Generally saved from a previous run of GridLMM on the same data.}

\item{save_V_folder}{Optional. A character vector giving a folder to save pre-processed V decomposition files for future / repeated use. 
If null, V decompositions are stored in memory}

\item{diagonalize}{If TRUE and the model includes only a single random effect, the "GEMMA" trick will be used to diagonalize V. This is done
by calculating the SVD of K, which can be slow for large samples.}

\item{svd_K}{If TRUE and \code{nrow(K) < nrow(Z)}, then the SVD is done on \eqn{K} instead of \eqn{ZKZ^T}}

\item{drop0_tol}{Values closer to zero than this will be set to zero when forming sparse matrices.}

\item{mc.cores}{Number of processor cores used for parallel evaluations. Note that if \code{clusterType = 'mclapply'}, the memory requires grow rapidly with \code{mc.cores}, because
the marker matrix gets duplicated in memory for each core.}

\item{clusterType}{`mclapply`}

\item{verbose}{Should progress be printed to the screen?}
}
\value{
A list with two elements:
\item{results}{A data frame with each row the results of the association test for a column of \code{X}, plus asssociated parameter values and statistics.}
\item{setup}{A list with several objects needed for re-running the model, including \code{V_setup} and \code{downdate_Xs}. 
  These can be re-passed to this function (or other GridLMM functions) to re-fit the model to the same data.}
}
\description{
Performs a GWAS using the adaptitve GridLMM algorithm. Can perform LRTs, Wald-tests, or calculate Bayes Factors.
}
\details{
Uses the adaptive GridLMM appraoch to evalue ML or REML solutions for each test, or calculate posterior distributions.
}
