% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GridLMM.R
\name{GridLMM_ML}
\alias{GridLMM_ML}
\title{Find the (Restricted) Maximim Likelihood solution to a linear mixed model by grid search}
\usage{
GridLMM_ML(
  formula,
  data,
  weights = NULL,
  relmat = NULL,
  initial_step = 0.5,
  tolerance = 0.01,
  ML = T,
  REML = T,
  V_setup = NULL,
  save_V_folder = NULL,
  diagonalize = T,
  mc.cores = my_detectCores(),
  verbose = T
)
}
\arguments{
\item{formula}{A two-sided linear formula as used in \code{\link[lme4]{lmer}} describing the fixed-effects
and random-effects of the model on the RHS and the response on the LHS. Note: correlated random-effects
are not implemented, so using one or two vertical bars (\code{|}) or one is identical. At least one random effect is needed.}

\item{data}{A data frame containing the variables named in \code{formula}.}

\item{weights}{An optional vector of observation-specific weights.}

\item{relmat}{A list of matrices that are proportional to the (within) covariance structures of the group level effects. 
The names of the matrices should correspond to the columns in \code{data} that are used as grouping factors. All levels
of the grouping factor should appear as rownames of the corresponding matrix.}

\item{initial_step}{See \strong{Details}.}

\item{tolerance}{See \strong{Details}.}

\item{ML}{If TRUE, ML solutions will be found.}

\item{REML}{If TRUE, REML solutions will be found.}

\item{V_setup}{Optional. A list produced by a GridLMM function containing the pre-processed V decompositions for each grid vertex, 
or the information necessary to create this. Generally saved from a previous run of GridLMM on the same data.}

\item{save_V_folder}{Optional. A character vector giving a folder to save pre-processed V decomposition files for future / repeated use. 
If null, V decompositions are stored in memory}

\item{diagonalize}{If TRUE and the model includes only a single random effect, the "GEMMA" trick will be used to diagonalize V. This is done
by calculating the SVD of K, which can be slow for large samples.}

\item{mc.cores}{Number of cores to use for parallel evaluations.}

\item{verbose}{Should progress be printed to the screen?}
}
\value{
A list with two elements:
\item{results}{A data frame with each row a ML or REML solution, and columns giving additional statistics and parameter values.}
\item{V_setup}{The \code{V_setup} object for this model. Can be re-passed to this function (or other GridLMM functions) to re-fit the model to the same data.}
}
\description{
Uses the GridLMM approach to find the ML or REML solution to variance componenents of a linear mixed model.
}
\details{
Note: this is not a particularly efficient technique for a single model. 
   However, the calculations for each grid vertex can be re-used in later fits (ex. GWAS).
   This function is used internally by other GridLMM functions to find starting values.
   
The function uses a hill-climbing heuristic on the grid. It starts be evaluating a grid with a step size of \code{initial_step} \eqn{h^2} units. 
    It then halves the step size and evaluates all grid vertices adjacent to the grid vertex with the highest ML (and/or REML) value.
    If a new maximum is found, the grid vertices adjacent to the new maximum are evaluated. This is repeated until no new maximum is found.
    At this point, if the step size is smaller than \code{tolerance}, the algorithm stops. Otherwise, the step size is halved again and the algorithm repeats.
}
