% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GridLMM.R
\name{GridLMM_posterior}
\alias{GridLMM_posterior}
\title{Evaluate the posterior of a linear mixed model using the Grid-sampling algorithm}
\usage{
GridLMM_posterior(
  formula,
  data,
  weights = NULL,
  relmat = NULL,
  h2_divisions = 10,
  h2_prior = function(h2s, n) 1/n,
  a = 0,
  b = 0,
  inv_prior_X = 0,
  target_prob = 0.99,
  thresh_nonzero = 10,
  thresh_nonzero_marginal = 0,
  V_setup = NULL,
  save_V_folder = NULL,
  diagonalize = T,
  mc.cores = my_detectCores(),
  verbose = T
)
}
\arguments{
\item{formula}{A two-sided linear formula as used in \code{\link[lme4]{lmer}} describing the fixed-effects
and random-effects of the model on the RHS and the response on the LHS. Note: correlated random-effects
are not implemented, so using one or two vertical bars (\code{|}) or one is identical. At least one random effect is needed.}

\item{data}{A data frame containing the variables named in \code{formula}.}

\item{weights}{An optional vector of observation-specific weights.}

\item{relmat}{A list of matrices that are proportional to the (within) covariance structures of the group level effects. 
The names of the matrices should correspond to the columns in \code{data} that are used as grouping factors. All levels
of the grouping factor should appear as rownames of the corresponding matrix.}

\item{h2_divisions}{Starting number of divisions of the grid for each variance component.}

\item{h2_prior}{Function that takes two arguments: 1) A vector of \code{h2s} (ie variance component proportions), and 
2) An integer giving the number of vertices in the full grid. The function should return a non-negative value giving the prior 
weight to the grid cell corresponding to \code{h2s}.}

\item{a, b}{Shape and Rate parameters of the Gamma prior for the residual variance \eqn{\sigma^2}. Setting both to zero gives a limiting "default" prior.}

\item{inv_prior_X}{Vector of values for the prior precision of each of the fixed effects (including an intercept). Will be recycled if necessary.}

\item{target_prob}{See \strong{Details}.}

\item{thresh_nonzero}{See \strong{Details}.}

\item{thresh_nonzero_marginal}{See \strong{Details}.}

\item{V_setup}{Optional. A list produced by a GridLMM function containing the pre-processed V decompositions for each grid vertex, 
or the information necessary to create this. Generally saved from a previous run of GridLMM on the same data.}

\item{save_V_folder}{Optional. A character vector giving a folder to save pre-processed V decomposition files for future / repeated use. 
If null, V decompositions are stored in memory}

\item{diagonalize}{If TRUE and the model includes only a single random effect, the "GEMMA" trick will be used to diagonalize V. This is done
by calculating the SVD of K, which can be slow for large samples.}

\item{mc.cores}{Number of cores to use for parallel evaluations.}

\item{verbose}{Should progress be printed to the screen?}

\item{svd_K}{If TRUE and \code{nrow(K) < nrow(Z)}, then the SVD is done on \eqn{K} instead of \eqn{ZKZ^T}}

\item{drop0_tol}{Values closer to zero than this will be set to zero when forming sparse matrices.}
}
\value{
A list with three elements:
\item{h2s_results}{A data frame with each row an evaluated grid vertex, with the first \eqn{l} columns giving the \eqn{h^2}'s, and the final column the 
    corresponding posterior mass}
\item{h2s_solutions}{A list with the parameters of the NIG distribution for each grid vertex}
\item{V_setup}{The \code{V_setup} object for this model. Can be re-passed to this function (or other GridLMM functions) to re-fit the model to the same data.}
}
\description{
Performs approximate posterior inference of a linear mixed model by evaluating the posterior over a grid
   of values of the variance component proportions. The grid evaluation uses a heuristic to avoid traversing 
   the entire grid. This should work well as long as the posterior is unimodal.
}
\details{
Posterior inference involves an adaptive grid search. Generally, we start with a very coarse grid (with as few as 2-3 vertices per variance component)
   and then progressively increase the grid resolution focusing only on regions of high posterior probability. This is controlled
   by \code{h2_divisions}, \code{target_prob}, \code{thresh_nonzero}, and \code{thresh_nonzero_matrginal}. The sampling algorithm is as follows:
   \itemize{
   \item Start by evaluating the posterior at each vertex of a trial grid with resolution \eqn{m}
   \item Find the minimum number of vertices needed to sum to \code{target_prob} of the current (discrete) posterior. 
      Repeat for the marginal posteriors of each variance component#'    
   \item If these numbers are smaller than \code{thresh_nonzero} or \code{thresh_nonzero_matrginal}, respectively, form a new grid
      by increasing the grid resolution to \eqn{m/2}. Otherwise, STOP.
   \item Begin evaluating the posterior at the new grid only at those grid vertices that are adjacent (in any dimension) to any of the top
      grid vertices in the old grid.
   \item Re-evaluate the distribution of the posterior over the new grid. If any new vertices contribute to the top \code{target_prob} fraction of the 
      overall posterior, include these in the "top" set and return to step 4. 
      Note - the prior weights for the grid vertices must be updated each time the grid increases in resolution.
   \item Repeat steps 4-5 until no new grid vertices contribute to the "top" set.
   \item Repeat steps 2-6 until a STOP is reached at step 3.
   }
   
Note: Default parameters for priors give flat (improper) priors. These should be used with care, especially for calculations of Bayes Factors.
}
