% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{my_detectCores}
\alias{my_detectCores}
\title{Detect number of available cores on system}
\usage{
my_detectCores()
}
\value{
ncores integer of number of available cores
}
\description{
\code{my_detectCores()} detects the number of cores available for calls to \code{\link{mclapply}}
or \code{\link{registerDoParallel}}.
}
\details{
This uses RcppParallel::defaultNumThreads() which seems to generally work for me.
}
\examples{

my_detectCores()
}
