% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{prep_LDAK_Kinship}
\alias{prep_LDAK_Kinship}
\title{Prepare a Kinship matrix for use with LDAK}
\usage{
prep_LDAK_Kinship(K, kinfile, LDAK_program = "LDAK/ldak5")
}
\arguments{
\item{K}{A matrix. Must have rownames.}

\item{kinfile}{The base of the filename fo the kinship files to be created, including path.}

\item{LDAK_program}{The path to the LDAK program}
}
\value{
A character with the base file name of the converted matrix.
}
\description{
Takes an R matrix and saves it to disk in a format that LDAK can use
}
\details{
The matrix is first saved as a text file (without row/column names), and then 
    LDAK is called by \code{system} with option "--convert-raw" to convert it into the format
    used by LDAK. Temporary files are removed. The .grn.id file contains the rownames of \code{K} repeated in two columns.
}
