% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{set_p_test}
\alias{set_p_test}
\title{Set the number of markers used in a GRM}
\usage{
set_p_test(V_setup, p_test = NULL, p = NULL)
}
\arguments{
\item{V_setup}{A list of setup variables, as returned by a GridLMM function}

\item{p_test}{A named vector with names corresponding to random effects in the model (as found in \code{V_setup$RE_setup}), 
and values equal to the typical number of markers to be \strong{removed} for a typical test}

\item{p}{A named vector as for \code{p_test} giving the total number of markers used to create the original GRM}
}
\value{
The modified \code{V_setup} list
}
\description{
For GRMs created as X'X, it can be useful to remove "proximal" markers from the GRM when testing a focal marker.
To do this in GridLMM, we need to know \code{p} (how many markers were in the original GRM), and \code{p_test},
how many markers are to be \strong{removed} for a typical test
}
