% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{my_detectCores}
\alias{my_detectCores}
\title{Detect number of available cores on system}
\usage{
my_detectCores()
}
\value{
ncores integer of number of avaialble cores
}
\description{
\code{my_detectCores()} detects the numer of cores available for calls to \code{\link{mclapply}}
or \code{\link{registerDoParallel}}.
}
\details{
The command is specific for clusters running slurm. 
Accesses the \code{SLURM_CPUS_PER_TASK} environmental variable. If this
variable doesn't exist it uses \code{\link[parallel]{detectCores}}.
}
\examples{

my_detectCores()
}
